Add Makefile + results for MetaMathQA task (#2593)
These are the first results for the MetaMathQA task and also the first
test of the Makefile used to run these tests.

The Makefile offers the functionality to run individual experiments by
specifying the result you want to have, e.g.
`make results/adalora--llama-3.2-3B-rank32[...].json`. Alternatively
you can simply run `make` for `make all` which runs all experiments
that don't have a result yet or which have outdated configs (comparing
result timestamp and config timestamp).

The results are from the main branch. No errors happened during the run.
There were errors with a compute instance that used a A10G 24GB because
of OOM. L40S with 48GB RAM was fine.


* Make sure to use original batch size for OFT

This was not done previously because of runner memory constraints.

* Remove timestamp from result files

We're tracking the results in git for now which makes
looking back easy enough (`git restore -s <rev> results`).
This makes it easier for `make` to track the results that
are already computed and which need to change.