Auto-tagging of PEFT models (#2599)
Features like inference need correctly set tags on the repo / the model card
in order to be available. Also the Hub uses tags to index the models and make
them searchable.

With this change PEFT tags models automatically as lora if they happen to
be trained with LoRA, the base model and a custom `peft:method:<the method>`
tag.

* Base model tags were never supported, they are now

Before PEFT simply ignored tags provided by the base model. Now the
base model tags are added to the PEFT-specific model tags.

* Tag 'tranformers' and add pipeline tag if possible

We remove the `peft:method:*` tag because this change needs more discussion
and is partially unrelated to this change. It is replaced by the necessary
`transformers` tag if the model is based on transformers.

We're also trying to resolve the pipeline tag automatically if it isn't set.
While there is the `transformers.pipelines.base.SUPPORTED_PEFT_TASKS` mapping
it is not sufficient to resolve the pipeline tag automatically since it is
not a 1:1 mapping. Only the causal LM case is a unique mapping.

---------

Co-authored-by: Benjamin Bossan <BenjaminBossan@users.noreply.github.com>