[WIP] Implement LoRA for nn.Parameter
Normally, nn.Parameter cannot be targeted with LoRA adapters. This can
be problematic, e.g. when there are MoE layers that use nn.Parameter
directly, or when there is nn.Linear but the weight is passed directly
instead of calling forward (e.g. MHA).

It would be possible to craft a solution involving a special LoRA layer
for each of the modules that use nn.Parameter directly (e.g. lora.MHA)
but that doesn't scale. This PR is an attempt at implementing a direct
way to target nn.Parameter making use of torch.nn.parametrize.

The current state of the PR is WIP, the next step is to add a
dispatching mechanism. This is not trivial, as we don't want the new
changes to accidentally affect the current matching logic. Probably the
best way is to add a completely new config variable (e.g.
target_parameters) that does not interfere with the current code.