[WIP] ENH: Adapter injection based on state_dict
Make it possible to inject the PEFT adapters based on a state_dict
instead of the PEFT config.

See https://github.com/huggingface/diffusers/issues/11874 for context.

Description

Right now, when creating a PEFT adapter like LoRA, the adapter layers
are injected based on the PEFT config, most notably the entries in
target_modules, but other arguments also play into this. Generally, this
is a good approach, but it breaks down in some situations. For instance,
in diffusers, we often have the situation that the checkpoint was
created without PEFT/diffusers, thus there is no PEFT config, only the
state_dict. To load these checkpoints in diffusers, the current approach
is to reverse-engineer a valid PEFT config based on the keys in the
state_dict.

Unfortunately, this is error prone. Moreover, not every combination of
state_dict keys can be easily expressed in a PEFT config through a
combination of target_modules, exclude_modules, etc. Yes, in theory
everything can be expressed by passing target_module=regex_pattern, but
reverse engineering such a regex is correctly and efficiently is very
hard (and thus currently not done).

This PR implements a completely different approach to inject adapters.
Instead of relying on the PEFT config to determine which layers to
target, it takes the state_dict directly as the source of truth. This
should allow to exactly match what is desired.

Implementation details

Technically, it is not necessary to pass the state_dict, we are only
interested in the keys. I still called the argument state_dict, since
that is typically what we have at this point, but this can be easily
changed.

I thought it might be a good idea, if the state_dict is used, to still
check what modules would have been targeted if we had used the PEFT
config. Then, the results are compared and a warning is given if they
differ. This allows the user to see if the PEFT config is not correctly
specified. While running some diffusers tests, I never encountered this
situation, though.

However, if we plan, for instance, to get rid of all the reverse
engineering of the PEFT config in diffusers, it would make more sense to
not give this warning.

Progress

There is still a lot to do when it comes to finishing the PR, most
notably expanding it to all methods, not just LoRA, adding
documentation, and adding tests.

To test this PR right now, install this branch directly. Then, inside of
diffusers, change this line:

https://github.com/huggingface/diffusers/blob/754fe85cace17ae8615d53578d6d842c9e4a1bd9/src/diffusers/loaders/peft.py#L321

- inject_adapter_in_model(lora_config, self, adapter_name=adapter_name, **peft_kwargs)
+ inject_adapter_in_model(lora_config, self, adapter_name=adapter_name, state_dict=state_dict, **peft_kwargs)

That's it. I tested the change with SD:

RUN_SLOW=1 pytest tests/lora/test_lora_layers_sd.py -v

and all tests passed. A new test based on the original issue of
https://github.com/huggingface/diffusers/issues/11874 should be added to
confirm that this PR solves this difficult case.