FIX Missing device map for facebook/opt-125m
Fixes the failing EETQ test in the nighly multi device CI.

In #2612, fixed device_maps were added for multi-GPU training as we
could not rely on device_map="auto". While doing this change, one
device_map was missing, namely for facebook/opt-125m, which is used in
the EETQ multi device test. This device_map was now added. This makes
the test pass locally.