FIX DynamicCache max_cache_len attribute error
Resolves current CI errors with prefix tuning.

Due to some recent changes in transformers (surfaced by
https://github.com/huggingface/transformers/pull/39797), checking
hasattr(cache, max_cache_len) results in an error:

>>> cache = DynamicCache()
>>> hasattr(cache, "max_cache_len")
Traceback (most recent call last):
  File "/home/name/work/forks/transformers/foo.py", line 9, in <module>
    hasattr(cache, "max_cache_len")
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/name/work/forks/transformers/src/transformers/cache_utils.py", line 916, in max_cache_len
    return max(values)
           ^^^^^^^^^^^
ValueError: max() iterable argument is empty

This has been reported and will be fixed in transformers. On the PEFT
side, it is safeest check the cache type and avoid accessing this
attribute in the first place, which is what this PR does.

Morever, that PR also changed the argument order to initialize
HybridCache (will probably also be reverted in transformers), which is
also taken into account in this PR by only using keyword arguments.