ENH Small speedups to adapter injection
See
https://github.com/huggingface/diffusers/issues/11816#issuecomment-3281290153

This PR implements two small improvements to the speed of adapter
injection. On a benchmark based on the linked issue, the first change
leads to a speedup of 21% and the second change of another 3%. It's not
that much, but as the changes don't make the code more complicated,
there is really no reason not to take them.

The optimizations don't add any functional change but are simply based
on not recomputing the same values multiple times. Therefore, unless I'm
missing something, they should strictly improve runtime.

Note:

Be careful when profiling this: Each operation is very quick but can be
perfomed millions of times. If the profiler adds overhead, it can
completely skew the results. E.g. with pyinstrument, just enabling the
profiler increases execution time (after optimization) from ~15 sec to
~21 sec.